+++ src/main/java/constants/Constants.java
@@ -2,6 +2,6 @@ package constants;
 
 public class Constants {
     public static final int LIST_SIZE = 100;
-    public static final String LINE_SEPARATOR = "____________________________________________________________";
+    public static final String LINE_SEPARATOR = "____________________________________________________________\n";
     public static final String INDENT = "%s";
 }
